feat(blueprints): Add minimum core version to blueprints. Move blueprint restore code to api/blueprints
